requirements.txt
httpx
websockets
click
bibtexparser
python-slugify
python-dotenv